penumbra_sdk_proto::cosmos::bank::v1beta1::query_server
Trait Query

pub trait Query:
    Send
    + Sync
    + 'static {
    // Required methods
    fn balance<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryBalanceRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryBalanceResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn all_balances<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryAllBalancesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryAllBalancesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn spendable_balances<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QuerySpendableBalancesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QuerySpendableBalancesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn spendable_balance_by_denom<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QuerySpendableBalanceByDenomRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QuerySpendableBalanceByDenomResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn total_supply<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryTotalSupplyRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryTotalSupplyResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn supply_of<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QuerySupplyOfRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QuerySupplyOfResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn params<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryParamsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryParamsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn denom_metadata<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryDenomMetadataRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomMetadataResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn denoms_metadata<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryDenomsMetadataRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomsMetadataResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn denom_owners<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryDenomOwnersRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomOwnersResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_enabled<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QuerySendEnabledRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QuerySendEnabledResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Available on crate feature rpc only.

Generated trait containing gRPC methods that should be implemented for use with QueryServer.

Required Methods

fn balance<'life0, 'async_trait>( &'life0 self, request: Request<QueryBalanceRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryBalanceResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Balance queries the balance of a single coin for a single account.

fn all_balances<'life0, 'async_trait>( &'life0 self, request: Request<QueryAllBalancesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryAllBalancesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

AllBalances queries the balance of all coins for a single account.

When called from another module, this query might consume a high amount of gas if the pagination field is incorrectly set.

fn spendable_balances<'life0, 'async_trait>( &'life0 self, request: Request<QuerySpendableBalancesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QuerySpendableBalancesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

SpendableBalances queries the spendable balance of all coins for a single account.

When called from another module, this query might consume a high amount of gas if the pagination field is incorrectly set.

Since: cosmos-sdk 0.46

fn spendable_balance_by_denom<'life0, 'async_trait>( &'life0 self, request: Request<QuerySpendableBalanceByDenomRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QuerySpendableBalanceByDenomResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

SpendableBalanceByDenom queries the spendable balance of a single denom for a single account.

When called from another module, this query might consume a high amount of gas if the pagination field is incorrectly set.

Since: cosmos-sdk 0.47

fn total_supply<'life0, 'async_trait>( &'life0 self, request: Request<QueryTotalSupplyRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryTotalSupplyResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

TotalSupply queries the total supply of all coins.

When called from another module, this query might consume a high amount of gas if the pagination field is incorrectly set.

fn supply_of<'life0, 'async_trait>( &'life0 self, request: Request<QuerySupplyOfRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QuerySupplyOfResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

SupplyOf queries the supply of a single coin.

When called from another module, this query might consume a high amount of gas if the pagination field is incorrectly set.

fn params<'life0, 'async_trait>( &'life0 self, request: Request<QueryParamsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryParamsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Params queries the parameters of x/bank module.

fn denom_metadata<'life0, 'async_trait>( &'life0 self, request: Request<QueryDenomMetadataRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomMetadataResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

DenomsMetadata queries the client metadata of a given coin denomination.

fn denoms_metadata<'life0, 'async_trait>( &'life0 self, request: Request<QueryDenomsMetadataRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomsMetadataResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

DenomsMetadata queries the client metadata for all registered coin denominations.

fn denom_owners<'life0, 'async_trait>( &'life0 self, request: Request<QueryDenomOwnersRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomOwnersResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

DenomOwners queries for all account addresses that own a particular token denomination.

When called from another module, this query might consume a high amount of gas if the pagination field is incorrectly set.

Since: cosmos-sdk 0.46

fn send_enabled<'life0, 'async_trait>( &'life0 self, request: Request<QuerySendEnabledRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QuerySendEnabledResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

SendEnabled queries for SendEnabled entries.

This query only returns denominations that have specific SendEnabled settings. Any denomination that does not have a specific setting will use the default params.default_send_enabled, and will not be returned by this query.

Since: cosmos-sdk 0.47