penumbra_sdk_proto::penumbra::core::component::auction
Module v1

Modules

event_dutch_auction_ended

Nested message and enum types in EventDutchAuctionEnded.

query_service_clientrpc

Generated client implementations.

query_service_serverrpc

Generated server implementations.

Structs

ActionDutchAuctionEnd

Terminate the auction associated with the specified auction_id

ActionDutchAuctionSchedule

Initiates a Dutch auction using protocol-controlled liquidity.

ActionDutchAuctionScheduleView

An ActionDutchAuctionSchedule augmented with additional metadata.

ActionDutchAuctionWithdraw

Withdraw funds from the ended auction associated with the specified auction_id

ActionDutchAuctionWithdrawPlan

A plan to a ActionDutchAuctionWithdraw which contains both private and public data.

ActionDutchAuctionWithdrawView

An ActionDutchAuctionWithdraw augmented with additional metadata.

AuctionId

A unique identifier for an auction, obtained from hashing a domain separator along with the immutable part of an auction description.

AuctionNft

A bearer NFT tracking ownership of an auction and its proceeds.

AuctionParameters

The configuration parameters for the auction component.

AuctionStateByIdRequest

AuctionStateByIdResponse

AuctionStateByIdsRequest

AuctionStateByIdsResponse

DutchAuction

DutchAuctionDescription

Describes a Dutch auction using programmatic liquidity on the DEX.

DutchAuctionState

EventDutchAuctionEnded

EventDutchAuctionScheduled

EventDutchAuctionUpdated

EventDutchAuctionWithdrawn

EventValueCircuitBreakerCredit

A message emitted when value flows into the auction component.

EventValueCircuitBreakerDebit

A message emitted when value flows out of the auction component.

GenesisContent

Genesis data for the auction component.