penumbra_sdk_proto::penumbra::core::component::dex
Module v1

Modules

position_state

Nested message and enum types in PositionState.

query_service_clientrpc

Generated client implementations.

query_service_serverrpc

Generated server implementations.

simulate_trade_request

Nested message and enum types in SimulateTradeRequest.

simulation_service_clientrpc

Generated client implementations.

simulation_service_serverrpc

Generated server implementations.

swap_claim_view

Nested message and enum types in SwapClaimView.

swap_execution

Nested message and enum types in SwapExecution.

swap_view

Nested message and enum types in SwapView.

Structs

ArbExecutionRequest

ArbExecutionResponse

ArbExecutionsRequest

ArbExecutionsResponse

BareTradingFunction

The minimum amount of data describing a trading function.

BatchSwapOutputData

Records the result of a batch swap on-chain.

BatchSwapOutputDataRequest

Requests batch swap data associated with a given height and trading pair from the view service.

BatchSwapOutputDataResponse

CandlestickData

CandlestickDataRequest

CandlestickDataResponse

CandlestickDataStreamRequest

CandlestickDataStreamResponse

DexParameters

DirectedTradingPair

Encodes a trading pair starting from asset start and ending on asset end.

EventArbExecution

EventBatchSwap

EventCandlestickData

Emitted whenever there’s non-empty candlestick data for a particular pair.

EventPositionClose

EventPositionExecution

EventPositionOpen

EventPositionWithdraw

EventQueuePositionClose

EventSwap

EventSwapClaim

EventValueCircuitBreakerCredit

Indicates that value was added to the DEX.

EventValueCircuitBreakerDebit

Indicates that value is leaving the DEX.

GenesisContent

LiquidityPositionByIdRequest

LiquidityPositionByIdResponse

LiquidityPositionsByIdRequest

LiquidityPositionsByIdResponse

LiquidityPositionsByPriceRequest

LiquidityPositionsByPriceResponse

LiquidityPositionsRequest

LiquidityPositionsResponse

LpNft

An LPNFT tracking both ownership and state of a position.

Position

Data identifying a position.

PositionClose

A transaction action that closes a position.

PositionId

A hash of a Position.

PositionOpen

A transaction action that opens a new position.

PositionRewardClaim

Deprecated.

PositionRewardClaimPlan

Deprecated.

PositionState

The state of a position.

PositionWithdraw

A transaction action that withdraws funds from a closed position.

PositionWithdrawPlan

Contains private and public data for withdrawing funds from a closed position.

Reserves

The reserves of a position.

SimulateTradeRequest

SimulateTradeResponse

SpreadRequest

SpreadResponse

Swap

A transaction action that submits a swap to the dex.

SwapBody

The authorized data of a Swap transaction.

SwapClaim

A transaction action that obtains assets previously confirmed via a Swap transaction. Does not include a spend authorization signature, as it is only capable of consuming the NFT from a Swap transaction.

SwapClaimBody

Encapsulates the authorized fields of the SwapClaim action, used in signing.

SwapClaimPlan

SwapClaimView

SwapExecution

Contains the entire execution of a particular swap.

SwapExecutionRequest

SwapExecutionResponse

SwapExecutionsRequest

SwapExecutionsResponse

SwapPayload

SwapPlaintext

SwapPlan

SwapView

TradingFunction

The trading function for a specific pair. For a pair (asset_1, asset_2), a trading function is defined by: phi(R) = p*R_1 + q*R_2 and gamma = 1 - fee. The trading function is frequently referred to as “phi”.

TradingPair

Holds two asset IDs. Ordering doesn’t reflect trading direction. Instead, we require asset_1 < asset_2 as field elements, to ensure a canonical representation of an unordered pair.

ZkSwapClaimProof

A Penumbra ZK swap claim proof.

ZkSwapProof

A Penumbra ZK swap proof.