penumbra_sdk_proto::penumbra::core::component::distributions
Module v1

Modules

distributions_service_clientrpc

Generated client implementations.

distributions_service_serverrpc

Generated server implementations.

Structs

CurrentLqtPoolSizeRequest

Request for retrieving the pool size of the current epoch from the chain state.

CurrentLqtPoolSizeResponse

Response containing the pool size for the current epoch.

DistributionsParameters

Distribution configuration data.

EventLqtPoolSizeIncrease

Event emitted when the size of the LQT pool increases.

GenesisContent

Genesis data for the distributions module.

LqtPoolSizeByEpochRequest

Request for retrieving the pool size at a specific epoch.

LqtPoolSizeByEpochResponse

Response containing the pool size at a specific epoch.