penumbra_sdk_proto::penumbra::core::component::funding::v1::funding_service_server
Trait FundingService

pub trait FundingService:
    Send
    + Sync
    + 'static {
    // Required method
    fn lqt_check_nullifier<'life0, 'async_trait>(
        &'life0 self,
        request: Request<LqtCheckNullifierRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<LqtCheckNullifierResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Available on crate feature rpc only.

Generated trait containing gRPC methods that should be implemented for use with FundingServiceServer.

Required Methods

fn lqt_check_nullifier<'life0, 'async_trait>( &'life0 self, request: Request<LqtCheckNullifierRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<LqtCheckNullifierResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Checks if a particular nullifier has already been used in the current epoch.