penumbra_sdk_proto::penumbra::core::component::governance
Module v1

Modules

delegator_vote_view

Nested message and enum types in DelegatorVoteView.

proposal

Nested message and enum types in Proposal.

proposal_outcome

Nested message and enum types in ProposalOutcome.

proposal_state

Nested message and enum types in ProposalState.

query_service_clientrpc

Generated client implementations.

query_service_serverrpc

Generated server implementations.

vote

Nested message and enum types in Vote.

Structs

AllTalliedDelegatorVotesForProposalRequest

AllTalliedDelegatorVotesForProposalResponse

ChangedAppParameters

DEPRECATED

ChangedAppParametersSet

DEPRECATED

CommunityPoolDeposit

CommunityPoolOutput

CommunityPoolSpend

DelegatorVote

DelegatorVoteBody

DelegatorVotePlan

DelegatorVoteView

EncodedParameter

EventDelegatorVote

EventProposalDepositClaim

EventProposalFailed

EventProposalPassed

EventProposalSlashed

EventProposalSubmit

EventProposalWithdraw

EventValidatorVote

GenesisContent

Governance genesis state.

GovernanceParameters

Governance configuration data.

NextProposalIdRequest

NextProposalIdResponse

Proposal

A proposal to be voted upon.

ProposalDataRequest

ProposalDataResponse

ProposalDepositClaim

ProposalInfoRequest

ProposalInfoResponse

ProposalListRequest

Requests the list of all proposals.

ProposalListResponse

The data for a single proposal.

ProposalOutcome

The outcome of a concluded proposal.

ProposalRateDataRequest

Requests the validator rate data for a proposal.

ProposalRateDataResponse

The rate data for a single validator.

ProposalState

The current state of a proposal.

ProposalSubmit

ProposalWithdraw

Ratio

The ratio between two numbers, used in governance to describe vote thresholds and quorums.

Tally

A tally of votes on a proposal.

ValidatorVote

ValidatorVoteBody

ValidatorVoteReason

ValidatorVotesRequest

Requests the list of all validator votes for a given proposal.

ValidatorVotesResponse

The data for a single validator vote.

Vote

A vote on a proposal.

VotingPowerAtProposalStartRequest

VotingPowerAtProposalStartResponse

ZkDelegatorVoteProof

A Penumbra ZK delegator vote proof.

Enums

ProposalKind

All the different kinds of proposals.