penumbra_sdk_proto::penumbra::core::component::sct
Module v1

Modules

commitment_source

Nested message and enum types in CommitmentSource.

query_service_clientrpc

Generated client implementations.

query_service_serverrpc

Generated server implementations.

Structs

AnchorByHeightRequest

AnchorByHeightResponse

CommitmentSource

Metadata describing the source of a commitment in the state commitment tree.

Epoch

An epoch is a sequentially numbered collection of contiguous blocks.

EpochByHeightRequest

EpochByHeightResponse

EventAnchor

Event recording an SCT anchor (global root).

EventBlockRoot

Event recording an SCT block root.

EventCommitment

Event recording a new commitment added to the SCT.

EventEpochRoot

Event recording an SCT epoch root.

GenesisContent

Sct-specific genesis content.

NullificationInfo

Records information about what transaction spent a nullifier.

Nullifier

SctFrontierRequest

SctFrontierResponse

SctParameters

Configuration data for the SCT component.

TimestampByHeightRequest

TimestampByHeightResponse