penumbra_sdk_proto::penumbra::core::component::shielded_pool
Module v1

Modules

event_outbound_fungible_token_refund

Nested message and enum types in EventOutboundFungibleTokenRefund.

fmd_meta_parameters

Nested message and enum types in FmdMetaParameters.

fmd_meta_parameters_algorithm_state

Nested message and enum types in FmdMetaParametersAlgorithmState.

genesis_content

Nested message and enum types in GenesisContent.

output_view

Nested message and enum types in OutputView.

query_service_clientrpc

Generated client implementations.

query_service_serverrpc

Generated server implementations.

spend_view

Nested message and enum types in SpendView.

Structs

AssetMetadataByIdRequest

Requests information on an asset by asset id

AssetMetadataByIdResponse

AssetMetadataByIdsRequest

Requests information on an asset by multiple asset ids

AssetMetadataByIdsResponse

EventBroadcastClue

ABCI Event recording a clue.

EventInboundFungibleTokenTransfer

EventOutboundFungibleTokenRefund

EventOutboundFungibleTokenTransfer

EventOutput

ABCI Event recording an output.

EventSpend

ABCI Event recording a spend.

FmdMetaParameters

The parameters which control how the FMD parameters evolve over time.

FmdMetaParametersAlgorithmState

Used to potentially store state for the FMD Meta Parameters algorithm.

FmdParameters

Parameters for Fuzzy Message Detection

FungibleTokenTransferPacketMetadata

Metadata about the packet associated with the transfer.

GenesisContent

Genesis data for the shielded pool component.

Note

NoteCiphertext

An encrypted note. 176 = 80(address) + 16(amount) + 32(asset ID) + 32(rseed) + 16(MAC) bytes.

NotePayload

The body of an output description, including only the minimal data required to scan and process the output.

NoteView

Output

Creates a new shielded note.

OutputBody

The body of an output description, containing only the effecting data describing changes to the ledger, and not the authorizing data that allows those changes to be performed.

OutputPlan

OutputView

ShieldedPoolParameters

Configuration data for the shielded pool component.

Spend

Spends a shielded note.

SpendBody

The body of a spend description, containing only the effecting data describing changes to the ledger, and not the authorizing data that allows those changes to be performed.

SpendPlan

SpendView

ZkNullifierDerivationProof

A Penumbra ZK nullifier derivation proof.

ZkOutputProof

A Penumbra ZK output proof.

ZkSpendProof

A Penumbra ZK spend proof.