penumbra_sdk_proto::penumbra::core::component::stake
Module v1

Modules

bonding_state

Nested message and enum types in BondingState.

funding_stream

Nested message and enum types in FundingStream.

query_service_clientrpc

Generated client implementations.

query_service_serverrpc

Generated server implementations.

validator_state

Nested message and enum types in ValidatorState.

Structs

BaseRateData

Describes the base reward and exchange rates in some epoch.

BondingState

Describes the unbonding state of a validator’s stake pool.

CurrentConsensusKeys

Tracks our view of Tendermint’s view of the validator set, so we can keep it from getting confused.

CurrentValidatorRateRequest

CurrentValidatorRateResponse

Delegate

A transaction action adding stake to a validator’s delegation pool.

DelegationChanges

A list of pending delegations and undelegations.

EventDelegate

EventRateDataChange

EventSlashingPenaltyApplied

Indicates a slashing penalty was applied to a validator’s reward rates.

EventTombstoneValidator

EventUndelegate

EventValidatorBondingStateChange

EventValidatorDefinitionUpload

EventValidatorMissedBlock

EventValidatorStateChange

EventValidatorVotingPowerChange

FundingStream

A portion of a validator’s commission.

GenesisContent

Genesis data for the staking component.

GetValidatorInfoRequest

Requests information about a specific validator.

GetValidatorInfoResponse

Penalty

Tracks slashing penalties applied to a validator in some epoch.

RateData

Describes the reward and exchange rates and voting power for a validator in some epoch.

StakeParameters

Staking configuration data.

Undelegate

A transaction action withdrawing stake from a validator’s delegation pool.

UndelegateClaim

A transaction action finishing an undelegation, converting (slashable) “unbonding tokens” to (unslashable) staking tokens.

UndelegateClaimBody

UndelegateClaimPlan

Uptime

Track’s a validator’s uptime.

Validator

Describes a validator’s configuration data.

ValidatorDefinition

A transaction action (re)defining a validator.

ValidatorInfo

Combines all validator info into a single packet.

ValidatorInfoRequest

Requests information on the chain’s validators.

ValidatorInfoResponse

ValidatorList

For storing the list of keys of known validators.

ValidatorPenaltyRequest

Requests the compounded penalty for a validator over a range of epochs.

ValidatorPenaltyResponse

ValidatorState

Describes the state of a validator

ValidatorStatus

Describes the current state of a validator on-chain

ValidatorStatusRequest

ValidatorStatusResponse

ValidatorUptimeRequest

ValidatorUptimeResponse

ZkUndelegateClaimProof

A Penumbra ZK undelegate claim proof.