penumbra_sdk_proto::penumbra::crypto::tct
Module v1

Structs

MerklePathChunk

A set of 3 sibling hashes in the auth path for some note commitment.

MerkleRoot

StateCommitment

StateCommitmentProof

An authentication path from a state commitment to the root of the state commitment tree.