penumbra_sdk_proto::penumbra::util::tendermint_proxy::v1::tendermint_proxy_service_server
Trait TendermintProxyService

pub trait TendermintProxyService:
    Send
    + Sync
    + 'static {
    // Required methods
    fn get_status<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetStatusRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetStatusResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn broadcast_tx_async<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BroadcastTxAsyncRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BroadcastTxAsyncResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn broadcast_tx_sync<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BroadcastTxSyncRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BroadcastTxSyncResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_tx<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTxRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetTxResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn abci_query<'life0, 'async_trait>(
        &'life0 self,
        request: Request<AbciQueryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<AbciQueryResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_by_height<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetBlockByHeightRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetBlockByHeightResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Available on crate feature rpc only.

Generated trait containing gRPC methods that should be implemented for use with TendermintProxyServiceServer.

Required Methods

fn get_status<'life0, 'async_trait>( &'life0 self, request: Request<GetStatusRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetStatusResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Status queries the current status.

fn broadcast_tx_async<'life0, 'async_trait>( &'life0 self, request: Request<BroadcastTxAsyncRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BroadcastTxAsyncResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Broadcast a transaction asynchronously.

fn broadcast_tx_sync<'life0, 'async_trait>( &'life0 self, request: Request<BroadcastTxSyncRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BroadcastTxSyncResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Broadcast a transaction synchronously.

fn get_tx<'life0, 'async_trait>( &'life0 self, request: Request<GetTxRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetTxResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Fetch a transaction by hash.

fn abci_query<'life0, 'async_trait>( &'life0 self, request: Request<AbciQueryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<AbciQueryResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

ABCIQuery defines a query handler that supports ABCI queries directly to the application, bypassing Tendermint completely. The ABCI query must contain a valid and supported path, including app, custom, p2p, and store.

fn get_block_by_height<'life0, 'async_trait>( &'life0 self, request: Request<GetBlockByHeightRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetBlockByHeightResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

GetBlockByHeight queries block for given height.