penumbra_sdk_proto::penumbra::view
Module v1

Modules

authorize_and_build_response

Nested message and enum types in AuthorizeAndBuildResponse.

broadcast_transaction_response

Nested message and enum types in BroadcastTransactionResponse.

delegations_by_address_index_request

Nested message and enum types in DelegationsByAddressIndexRequest.

transaction_planner_request

Nested message and enum types in TransactionPlannerRequest.

unbonding_tokens_by_address_index_request

Nested message and enum types in UnbondingTokensByAddressIndexRequest.

view_service_clientrpc

Generated client implementations.

view_service_serverrpc

Generated server implementations.

witness_and_build_response

Nested message and enum types in WitnessAndBuildResponse.

Structs

AddressByIndexRequest

AddressByIndexResponse

AppParametersRequest

Requests the current app parameters from the view service.

AppParametersResponse

AssetMetadataByIdRequest

Requests information on an asset by asset id

AssetMetadataByIdResponse

AssetsRequest

Requests all assets known to the view service.

AssetsResponse

Requests all assets known to the view service.

AuctionsRequest

Filters in an AuctionsRequest will be combined using AND logic – that is, the more filters you add, the fewer responses you’re likely to get.

AuctionsResponse

AuthorizeAndBuildRequest

AuthorizeAndBuildResponse

BalancesRequest

BalancesResponse

BroadcastTransactionRequest

BroadcastTransactionResponse

DelegationsByAddressIndexRequest

Requests ValueViews of delegation tokens for the given address index. The returned ValueViews will include the ValidatorInfo for the delegated validator in their extended_metadata fields.

DelegationsByAddressIndexResponse

Contains a ValueView of delegation tokens for the requested address index. The ValueView includes the ValidatorInfo for the delegated validator in cits extended_metadata field.

EphemeralAddressRequest

EphemeralAddressResponse

FmdParametersRequest

Requests the current FMD parameters from the view service.

FmdParametersResponse

GasPricesRequest

Requests the current gas prices from the view service.

GasPricesResponse

IndexByAddressRequest

IndexByAddressResponse

LatestSwapsRequest

Requests the latest swaps controlled by the user’s wallet.

LatestSwapsResponse

NoteByCommitmentRequest

NoteByCommitmentResponse

NotesForVotingRequest

A query for notes to be used for voting on a proposal.

NotesForVotingResponse

NotesRequest

A query for notes known by the view service.

NotesResponse

NullifierStatusRequest

NullifierStatusResponse

OwnedPositionIdsRequest

OwnedPositionIdsResponse

SpendableNoteRecord

A note plaintext with associated metadata about its status.

StatusRequest

Requests sync status of the view service.

StatusResponse

Returns the status of the view service and whether it is synchronized with the chain state.

StatusStreamRequest

Requests streaming updates on the sync height until the view service is synchronized.

StatusStreamResponse

A streaming sync status update

SwapByCommitmentRequest

SwapByCommitmentResponse

SwapRecord

TransactionInfo

TransactionInfoByHashRequest

TransactionInfoByHashResponse

TransactionInfoRequest

TransactionInfoResponse

TransactionPlannerRequest

TransactionPlannerResponse

TransparentAddressRequest

There’s only one transparent address per wallet, so this request has no parameters; the message exists to satisfy forward-compatibility properties.

TransparentAddressResponse

UnbondingTokensByAddressIndexRequest

Requests unbonding tokens for a given address index, with optional filtering for whether the tokens are currently claimable.

UnbondingTokensByAddressIndexResponse

Returns unbonding tokens for the given address index, optionally filtered by whether the tokens are currently claimable.

UnclaimedSwapsRequest

UnclaimedSwapsResponse

WalletIdRequest

WalletIdResponse

WitnessAndBuildRequest

WitnessAndBuildResponse

WitnessRequest

WitnessResponse