penumbra_sdk_proto::penumbra::view::v1
Module transaction_planner_request

Nested message and enum types in TransactionPlannerRequest.

Structs

ActionDutchAuctionEnd

ActionDutchAuctionSchedule

ActionDutchAuctionWithdraw

ActionLiquidityTournamentVote

Delegate

DelegatorVote

Output

Request message subtypes

PositionClose

PositionOpen

PositionWithdraw

Spend

Swap

SwapClaim

Undelegate

UndelegateClaim

Enums

FeeMode

Specifies either that the planner should compute fees automatically or that it should use a fixed fee amount.