penumbra_sdk_proto
Trait DomainType

pub trait DomainType
where
    Self: Clone + Sized + TryFrom<Self::Proto>,
    Self::Proto: Name + Message + Default + From<Self> + Send + Sync + 'static,
    Error: From<<Self as TryFrom<Self::Proto>>::Error>,
{
    type Proto;

    // Provided methods
    fn encode_to_vec(&self) -> Vec<u8>  { ... }
    fn to_proto(&self) -> Self::Proto { ... }
    fn decode<B: Buf>(buf: B) -> Result<Self> { ... }
}

A marker type that captures the relationships between a domain type (Self) and a protobuf type (Self::Proto).

Required Associated Types

type Proto

Provided Methods

fn encode_to_vec(&self) -> Vec<u8>

Encode this domain type to a byte vector, via proto type P.

fn to_proto(&self) -> Self::Proto

Convert this domain type to the associated proto type.

This uses the From impl internally, so it works exactly like .into(), but does not require type inference.

fn decode<B: Buf>(buf: B) -> Result<Self>

Decode this domain type from a byte buffer, via proto type P.

Object Safety

This trait is not object safe.