penumbra_sdk_proto::penumbra::core::component::funding
Module v1

Modules

action_liquidity_tournament_vote_view

Nested message and enum types in ActionLiquidityTournamentVoteView.

funding_parameters

Nested message and enum types in FundingParameters.

funding_service_clientrpc

Generated client implementations.

funding_service_serverrpc

Generated server implementations.

Structs

ActionLiquidityTournamentVote

An action for voting in a liquidity tournament.

ActionLiquidityTournamentVotePlan

The plan associated with a ActionLiquidityTournamentVote.

ActionLiquidityTournamentVoteView

EventFundingStreamReward

Indicates that a funding stream reward was paid.

EventLqtDelegatorReward

EventLqtPositionReward

EventLqtVote

FundingParameters

Funding component configuration data.

GenesisContent

Genesis data for the funding component.

LiquidityTournamentVoteBody

LqtCheckNullifierRequest

Request if a nullifier is already used in a specific epoch.

LqtCheckNullifierResponse

Response indicating whether the nullifier was already used.

ZkLiquidityTournamentVoteProof

A proof of the validity of a liquidity vote, wrt private state.