penumbra_sdk_proto::penumbra::core::transaction
Module v1

Modules

action

Nested message and enum types in Action.

action_plan

Nested message and enum types in ActionPlan.

action_view

Nested message and enum types in ActionView.

memo_view

Nested message and enum types in MemoView.

transaction_perspective

Nested message and enum types in TransactionPerspective.

transaction_summary

Nested message and enum types in TransactionSummary.

Structs

Action

A state change performed by a transaction.

ActionPlan

Describes a planned transaction action.

ActionView

A view of a specific state change action performed by a transaction.

AuthorizationData

The data required to authorize a transaction plan.

CluePlan

Describes a plan for forming a Clue.

DetectionData

Detection data used by a detection server performing Fuzzy Message Detection.

DetectionDataPlan

MemoCiphertext

The encrypted memo data describing information about the purpose of a transaction.

MemoPlaintext

The plaintext describing information about the purpose of a transaction.

MemoPlaintextView

MemoPlan

Describes a plan for forming the transaction memo.

MemoView

NullifierWithNote

PayloadKeyWithCommitment

Transaction

A Penumbra transaction.

TransactionBody

The body of a transaction.

TransactionBodyView

TransactionParameters

The parameters determining if a transaction should be accepted by the chain.

TransactionPerspective

A transaction perspective is a bundle of key material and commitment openings that allow generating a view of a transaction from that perspective.

TransactionPlan

Describes a planned transaction. Permits clients to prepare a transaction prior submission, so that a user can review it prior to authorizing its execution.

TransactionSummary

Represents a transaction summary containing multiple effects.

TransactionView

View of a Penumbra transaction.

WitnessData

The data required for proving when building a transaction from a plan.