penumbra_sdk_proto::penumbra::custody::threshold
Module v1

Modules

coordinator_round1

Nested message and enum types in CoordinatorRound1.

coordinator_round2

Nested message and enum types in CoordinatorRound2.

dkg_round2

Nested message and enum types in DKGRound2.

follower_round1

Nested message and enum types in FollowerRound1.

follower_round2

Nested message and enum types in FollowerRound2.

Structs

CoordinatorRound1

The message the coordinator sends in round 1 of the signing protocol.

CoordinatorRound2

The message the coordinator sends in round 2 of the signing protocol.

DkgRound1

The first message we broadcast in the DKG protocol.

DkgRound2

The second message we broadcast in the DKG protocol.

FollowerRound1

The first message the followers send back to the coordinator when signing.

FollowerRound2

The second message the followers send back to the coordinator when signing.

Signature

A signature proving that a message was created by the owner of a verification key.

VerificationKey

A key one can use to verify signatures.