penumbra_sdk_proto
Module serializers

Helper methods used for shaping the JSON (and other Serde) formats derived from the protos. TODO: Extract the Bech32 prefixes into a module at the root of the crate, and then remove this module entirely. We don’t want to be custom-shaping serde formats, we want to have those match protobuf JSON exactly.

Modules

bech32str

Bech32 serializers. Because Bech32 is parameterized by the HRP, this module implements (internal) helper functions that are used in submodules that fill in parameters for various key types.